+++ test/chart/data_pool.dart
@@ -1182,9 +1182,17 @@ List<LineTooltipItem?> lineChartGetTooltipItems(List<LineBarSpot> list) {
   return list.map((s) => lineTooltipItem1).toList();
 }
 
+Color lineChartGetGreenColor(LineBarSpot touchedSpot) {
+  return Colors.green;
+}
+
+Color lineChartGetRedColor(LineBarSpot touchedSpot) {
+  return Colors.red;
+}
+
 const LineTouchTooltipData lineTouchTooltipData1 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1194,7 +1202,7 @@ const LineTouchTooltipData lineTouchTooltipData1 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData1Clone = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1205,7 +1213,7 @@ const LineTouchTooltipData lineTouchTooltipData1Clone = LineTouchTooltipData(
 
 const LineTouchTooltipData lineTouchTooltipData2 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.red,
+  getTooltipColor: lineChartGetRedColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1215,7 +1223,7 @@ const LineTouchTooltipData lineTouchTooltipData2 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData3 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.2),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1226,7 +1234,7 @@ const LineTouchTooltipData lineTouchTooltipData3 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData4 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 13,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1237,7 +1245,7 @@ const LineTouchTooltipData lineTouchTooltipData4 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData5 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1248,7 +1256,7 @@ const LineTouchTooltipData lineTouchTooltipData5 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData6 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1260,7 +1268,7 @@ const LineTouchTooltipData lineTouchTooltipData6 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData7 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -2235,6 +2243,14 @@ ScatterTooltipItem? scatterChartGetTooltipItems(ScatterSpot spots) {
   );
 }
 
+Color scatterChartGetTooltipGreenColor(ScatterSpot spots) {
+  return Colors.green; //Color
+}
+
+Color scatterChartGetTooltipRedColor(ScatterSpot spots) {
+  return Colors.red; //Color
+}
+
 final ScatterSpot scatterSpot1 = ScatterSpot(1, 40);
 final ScatterSpot scatterSpot1Clone = ScatterSpot(1, 40);
 final ScatterSpot scatterSpot2 = ScatterSpot(-4, -8);
@@ -2297,7 +2313,7 @@ final ScatterChartData scatterChartData1 = ScatterChartData(
       fitInsideHorizontally: true,
       fitInsideVertically: false,
       maxContentWidth: 33,
-      tooltipBgColor: Colors.white,
+      getTooltipColor: (touchedSpot) => Colors.white,
       tooltipPadding: const EdgeInsets.all(23),
       tooltipRoundedRadius: 534,
     ),
@@ -2336,7 +2352,7 @@ final ScatterTouchTooltipData scatterTouchTooltipData1 =
     ScatterTouchTooltipData(
   tooltipRoundedRadius: 23,
   tooltipPadding: const EdgeInsets.all(11),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: scatterChartGetTooltipGreenColor,
   maxContentWidth: 33,
   fitInsideVertically: true,
   fitInsideHorizontally: false,
@@ -2347,7 +2363,7 @@ final ScatterTouchTooltipData scatterTouchTooltipData1Clone =
     ScatterTouchTooltipData(
   tooltipRoundedRadius: 23,
   tooltipPadding: const EdgeInsets.all(11),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: scatterChartGetTooltipGreenColor,
   maxContentWidth: 33,
   fitInsideVertically: true,
   fitInsideHorizontally: false,
@@ -2358,7 +2374,7 @@ final ScatterTouchTooltipData scatterTouchTooltipData2 =
     ScatterTouchTooltipData(
   tooltipRoundedRadius: 23,
   tooltipPadding: const EdgeInsets.all(11),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: scatterChartGetTooltipGreenColor,
   maxContentWidth: 33,
   fitInsideVertically: true,
   fitInsideHorizontally: false,
@@ -2370,7 +2386,7 @@ final ScatterTouchTooltipData scatterTouchTooltipData3 =
     ScatterTouchTooltipData(
   tooltipRoundedRadius: 23,
   tooltipPadding: const EdgeInsets.all(11),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: scatterChartGetTooltipGreenColor,
   maxContentWidth: 33,
   fitInsideVertically: true,
   fitInsideHorizontally: false,
@@ -2744,12 +2760,24 @@ BarTooltipItem getTooltipItem(
   return BarTooltipItem(rod.toY.toString(), textStyle);
 }
 
+Color getTooltipGreenColor(
+  BarChartGroupData group,
+) {
+  return Colors.green;
+}
+
+Color getTooltipBlueColor(
+  BarChartGroupData group,
+) {
+  return Colors.blue;
+}
+
 final BarTouchTooltipData barTouchTooltipData1 = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2760,7 +2788,7 @@ final BarTouchTooltipData barTouchTooltipData1Clone = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2771,7 +2799,7 @@ final BarTouchTooltipData barTouchTooltipData2 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2783,7 +2811,7 @@ final BarTouchTooltipData barTouchTooltipData3 = BarTouchTooltipData(
   fitInsideVertically: true,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2795,7 +2823,7 @@ final BarTouchTooltipData barTouchTooltipData4 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: false,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2807,7 +2835,7 @@ final BarTouchTooltipData barTouchTooltipData5 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23.00001,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2820,7 +2848,7 @@ final BarTouchTooltipData barTouchTooltipData6 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.blue,
+  getTooltipColor: getTooltipBlueColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2833,7 +2861,7 @@ final BarTouchTooltipData barTouchTooltipData7 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
   tooltipBorder: const BorderSide(color: Colors.red),
@@ -2845,7 +2873,7 @@ final BarTouchTooltipData barTouchTooltipData8 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   tooltipMargin: 12,
   tooltipBorder: const BorderSide(color: Colors.red),
@@ -2855,7 +2883,7 @@ final BarTouchTooltipData barTouchTooltipData9 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 333,
@@ -2866,7 +2894,7 @@ final BarTouchTooltipData barTouchTooltipData10 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2877,7 +2905,7 @@ final BarTouchTooltipData barTouchTooltipData11 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,